requirements.txt
grpcio==1.33.2
grpcio-tools==1.33.2
protobuf==3.13.0
six==1.15.0
setuptools~=50.3.2
tox==3.20.1
twine==3.2.0
pytest==6.0.1
pytest-cov==2.10.1
flask==1.1.2
waitress==1.4.4